(fix) Improve performance of patient chart
