+++ packages/esm-patient-chart-app/src/root.component.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useMemo } from 'react';
+import React from 'react';
 import { SWRConfig } from 'swr';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { WorkspaceWindowSizeProvider } from '@openmrs/esm-patient-common-lib';
-import { setLeftNav, unsetLeftNav, usePatient } from '@openmrs/esm-framework';
-import { dashboardPath, spaRoot, spaBasePath, basePath } from './constants';
+import { dashboardPath, spaRoot, basePath } from './constants';
 import WorkspaceWindow from './workspace/workspace-window.component';
 import PatientChart from './patient-chart/patient-chart.component';
 import SideMenu from './side-nav/side-menu.component';
@@ -16,13 +15,6 @@ const swrConfiguration = {
 };
 
 export default function Root() {
-  const { patientUuid } = usePatient();
-  const leftNavBasePath = useMemo(() => spaBasePath.replace(':patientUuid', patientUuid), [patientUuid]);
-  useEffect(() => {
-    setLeftNav({ name: 'patient-chart-dashboard-slot', basePath: leftNavBasePath });
-    return () => unsetLeftNav('patient-chart-dashboard-slot');
-  }, [leftNavBasePath]);
-
   return (
     <SWRConfig value={swrConfiguration}>
       <BrowserRouter basename={spaRoot}>
@@ -34,7 +26,7 @@ export default function Root() {
               <Route path={basePath} element={<PatientChart />} />
               <Route path={dashboardPath} element={<PatientChart />} />
             </Routes>
-            <WorkspaceWindow patientUuid={patientUuid} />
+            <WorkspaceWindow />
           </div>
         </WorkspaceWindowSizeProvider>
       </BrowserRouter>
