+++ packages/esm-patient-common-lib/src/index.ts
@@ -15,5 +15,4 @@ export * from './form-entry/form-entry';
 export * from './workspaces';
 export * from './programs/usePatientProgramEnrollment';
 export * from './offline/visit';
-export * from './offline/patient';
 export * from './patient-summary/patient-summary-extension-order';
